+++ packages/core/demo/demo-data/bar.ts
@@ -8,7 +8,6 @@ export const groupedBarData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				65000,
 				-29123,
@@ -19,7 +18,6 @@ export const groupedBarData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				32432,
 				-21312,
@@ -30,7 +28,6 @@ export const groupedBarData = {
 		},
 		{
 			label: "Dataset 3",
-			fillColors: [colors[2]],
 			data: [
 				-12312,
 				23232,
@@ -41,7 +38,6 @@ export const groupedBarData = {
 		},
 		{
 			label: "Dataset 4",
-			fillColors: [colors[3]],
 			data: [
 				-32423,
 				21313,
@@ -62,9 +58,6 @@ export const groupedBarOptions = {
 		bottom: {
 			scaleType: "labels",
 			secondary: true,
-		},
-		top: {
-			scaleType: "labels",
 		}
 	}
 };
@@ -92,7 +85,6 @@ export const simpleBarData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: colors,
 			data: [
 				65000,
 				29123,
@@ -139,7 +131,6 @@ export const simpleBarTimeSeriesData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: colors,
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -205,7 +196,6 @@ export const stackedBarData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				65000,
 				29123,
@@ -216,7 +206,6 @@ export const stackedBarData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				32432,
 				21312,
@@ -227,7 +216,6 @@ export const stackedBarData = {
 		},
 		{
 			label: "Dataset 3",
-			fillColors: [colors[2]],
 			data: [
 				12312,
 				23232,
@@ -238,7 +226,6 @@ export const stackedBarData = {
 		},
 		{
 			label: "Dataset 4",
-			fillColors: [colors[3]],
 			data: [
 				32423,
 				21313,
@@ -287,7 +274,6 @@ export const stackedBarTimeSeriesData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -313,7 +299,6 @@ export const stackedBarTimeSeriesData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				{
 					date: new Date(2019, 0, 3),
@@ -339,7 +324,6 @@ export const stackedBarTimeSeriesData = {
 		},
 		{
 			label: "Dataset 3",
-			fillColors: [colors[2]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -365,7 +349,6 @@ export const stackedBarTimeSeriesData = {
 		},
 		{
 			label: "Dataset 4",
-			fillColors: [colors[3]],
 			data: [
 				{
 					date: new Date(2019, 0, 2),