+++ packages/core/src/chart.ts
@@ -14,7 +14,8 @@ import { Component,
 	Title,
 	Legend,
 	LayoutComponent,
-	Tooltip
+	Tooltip,
+	Spacer
 } from "./components";
 import { Tools } from "./tools";
 
@@ -22,7 +23,6 @@ import { Tools } from "./tools";
 import {
 	DOMUtils,
 	Events,
-	Themes,
 	Transitions
 } from "./services/index";
 
@@ -31,8 +31,7 @@ export class Chart {
 	services: any = {
 		domUtils: DOMUtils,
 		events: Events,
-		transitions: Transitions,
-		themes: Themes
+		transitions: Transitions
 	};
 	model: ChartModel = new ChartModel(this.services);
 
@@ -81,6 +80,12 @@ export class Chart {
 		if (!this.components) {
 			return;
 		}
+
+		// Update all services
+		Object.keys(this.services).forEach(serviceName => {
+			const serviceObj = this.services[serviceName];
+			serviceObj.update();
+		});
 
 		// Render all components
 		this.components.forEach(component => component.render(animate));
@@ -164,6 +169,17 @@ export class Chart {
 			fullFrameComponentDirection = LayoutDirection.COLUMN_REVERSE;
 		}
 
+		const legendSpacerComponent = {
+			id: "spacer",
+			components: [
+				new Spacer(this.model, this.services)
+			],
+			growth: {
+				x: LayoutGrowth.PREFERRED,
+				y: LayoutGrowth.FIXED
+			}
+		};
+
 		const fullFrameComponent = {
 			id: "full-frame",
 			components: [
@@ -172,6 +188,7 @@ export class Chart {
 					this.services,
 					[
 						legendComponent,
+						legendSpacerComponent,
 						graphFrameComponent
 					],
 					{
@@ -189,6 +206,19 @@ export class Chart {
 		const topLevelLayoutComponents = [];
 		if (this.model.getOptions().title) {
 			topLevelLayoutComponents.push(titleComponent);
+
+			const titleSpacerComponent = {
+				id: "spacer",
+				components: [
+					new Spacer(this.model, this.services)
+				],
+				growth: {
+					x: LayoutGrowth.PREFERRED,
+					y: LayoutGrowth.FIXED
+				}
+			};
+
+			topLevelLayoutComponents.push(titleSpacerComponent);
 		}
 		topLevelLayoutComponents.push(fullFrameComponent);