+++ packages/core/src/components/graphs/bar-grouped.ts
@@ -1,11 +1,11 @@
 // Internal Imports
 import { Bar } from "./bar";
+import { BarOrientationOptions, Roles, TooltipTypes } from "../../interfaces";
 
 // D3 Imports
 import { select } from "d3-selection";
 import { color } from "d3-color";
 import { ScaleBand, scaleBand } from "d3-scale";
-import { TooltipTypes, BarOrientationOptions, ScaleTypes } from "../../interfaces";
 
 export class GroupedBar extends Bar {
 	type = "grouped-bar";
@@ -76,7 +76,9 @@ export class GroupedBar extends Bar {
 		// Add the bar groups that need to be introduced
 		const barGroupsEnter = barGroups.enter()
 			.append("g")
-			.classed("bars", true);
+				.classed("bars", true)
+				.attr("role", Roles.GROUP)
+				.attr("aria-labelledby", d => d);
 
 		// Update data on all bars
 		const bars = barGroupsEnter.merge(barGroups)
@@ -116,7 +118,11 @@ export class GroupedBar extends Bar {
 					return Math.abs(this.services.axes.getYValue(d, i) - this.services.axes.getYValue(0));
 				})
 				.attr("fill", d => this.model.getFillScale()[d.datasetLabel](d.label))
-				.attr("opacity", 1);
+				.attr("opacity", 1)
+				// a11y
+				.attr("role", Roles.GRAPHICS_SYMBOL)
+				.attr("aria-roledescription", "bar")
+				.attr("aria-label", d => d.value);
 		} else {
 			// code for horizontal orientation grouped bar
 			barsEnter.merge(bars)
@@ -132,7 +138,11 @@ export class GroupedBar extends Bar {
 					return this.groupScale(d.datasetLabel); })
 				.attr("height",  this.getBarWidth.bind(this))
 				.attr("fill", d => this.model.getFillScale()[d.datasetLabel](d.label))
-				.attr("opacity", 1);
+				.attr("opacity", 1)
+				// a11y
+				.attr("role", Roles.GRAPHICS_SYMBOL)
+				.attr("aria-roledescription", "bar")
+				.attr("aria-label", d => d.value);
 		}
 
 		// Add event listeners to elements drawn
@@ -146,7 +156,7 @@ export class GroupedBar extends Bar {
 		return datasets.map(dataset => ({
 			label: d,
 			datasetLabel: dataset.label,
-			value: dataset.data[index]
+			value: dataset.data[index].value ? dataset.data[index].value : dataset.data[index]
 		}));
 	}