+++ packages/core/src/interfaces/charts.ts
@@ -7,10 +7,6 @@ import { ChartTheme, BarOrientationOptions } from "./enums";
  */
 export interface BaseChartOptions {
 	/**
-	 * boolean to enable accessibility mode
-	 */
-	accessibility?: boolean;
-	/**
 	 * boolean to disable animations (enabled by default)
 	 */
 	animations?: boolean;
@@ -21,11 +17,11 @@ export interface BaseChartOptions {
 	/**
 	 * Optionally specify a width for the chart
 	 */
-	width?: number | string;
+	width?: string;
 	/**
 	 * Optionally specify a height for the chart
 	 */
-	height?: number | string;
+	height?: string;
 	/**
 	 * Optional function to generate the fill color based on datasetLabel, label, and/or value
 	 */
@@ -130,6 +126,9 @@ export interface PieChartOptions extends BaseChartOptions {
 			offsetY?: number;
 			horizontalLineLength?: number;
 			textMargin?: number;
+		},
+		labels?: {
+			formatter?: Function;
 		}
 	};
 }
@@ -145,6 +144,7 @@ export interface DonutChartOptions extends PieChartOptions {
 			numberFontSize?: Function;
 			titleFontSize?: Function;
 			titleYPosition?: Function;
+			numberFormatter?: Function;
 		};
 	};
 }